requirements.txt
flask==2.0.3
flask-sqlalchemy==2.5.1
sqlalchemy==1.4.46
werkzeug==2.0.3
python-dotenv==0.19.2